Clean up execution epics
